This commit was manufactured by cvs2svn to create tag 'Ext-4-2-7-rc2'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-4-2-7-rc2@14780 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/util_funcs.h
@@ -42,8 +42,8 @@ int  Search_Table( mib_table_t, void*, int);
 int  Add_Entry( mib_table_t, void*);
 void *Retrieve_Table_Data( mib_table_t, int*);
 
-int marker_uptime( marker_t pm );
-int marker_tticks( marker_t pm );
-int timeval_uptime( struct timeval *tv );
-int timeval_tticks( struct timeval *tv );
+u_long marker_uptime( marker_t pm );
+u_long marker_tticks( marker_t pm );
+u_long timeval_uptime( struct timeval *tv );
+u_long timeval_tticks( struct timeval *tv );
 #endif /* _MIBGROUP_UTIL_FUNCS_H */
